Scala: Support metavars as elements in interpolated strings (#11267)
When looking for patterns in interpolated strings like

```
- pattern: s"..."
```
it is possible to determine if your tainted user input goes into the
interpolated string.
But what is not possible is something like this:
```
- focus-metavariable: $TAINTED
- pattern: s"$BAD...$TAINTED"
- metavariable-regex:
      metavariable: $BAD
      regex: This text cannot be followed by a tainted value
```

In this PR:

- The lexer does not ignore $ inside an interpolated string when it is
followed by a metavar identifier
- In the sgrep mode (i.e., when parsing patterns) the parser
distinguishes between variables and metavariables by checking if the
identifier starts with $. In non-sgrep mode (i.e., when parsing
targets), the parser treats all identifiers as variables.