chore(lfs): create new resolution error type for inaccessible resources (semgrep/semgrep-proprietary#4994)
A common error class for us in lockfileless is inaccessible resources
that are
needed for the build. This is typically caused by incomplete private
package
manager configuration, but could theoretically also be caused by a
package
manager misconfiguration. I believe that in all cases, there is not much
we can
do on our end. 

So this PR begins to capture a common subset of the relevant errors;
those where a Maven parent POM could not be resolved. We identify these
errors
simply by matching on text that I observed in the Maven  output.

synced from Pro 2c8566b59c55e0f90db416f3948c3b6e38d4d9f7