Revert "Faster subproject discovery in Supply Chain scans" (semgrep/semgrep-proprietary#5126)
Some local testing shows cases where we drop subprojects as a result of
this change.

We've identified the change in behavior is related to `.gitignore`.
Previously, we ignored `.gitignore` when resolving subprojects. With
this commit, we respect `.gitignore` and so we don't scan ignored
subproject files.

Ultimately this is probably the correct behavior, as it matches how code
works and our documentation (which says we respect `.gitignore` files).
But this might cause issues for customers currently relying on this
behavior to generate lockfiles in CI that they don't want to check in
for us to scan, so we are going to revert it anyways and re-assess the
impact with customers.

Example repo (with vulnerable dependencies, do not install the packages
in the lockfiles in this repo!):


[hakmiller-scrat-tests-vulnerable.zip](https://github.com/user-attachments/files/23940754/hakmiller-scrat-tests-vulnerable.zip)


Reverts semgrep/semgrep-proprietary#5065

synced from Pro 660145f4324ee1ff729a1ea514d8dd2f9a25d996