fix(telemetry): Propagate opentelemetry context through threads (semgrep/semgrep-proprietary#5205)
By default, opentelemetry contexts are not propagated to new threads in Python. Because we perform rule validation in worker threads, this meant that spans from that work were not connected to the main process in any way. When viewing a trace, they would appear to be missing.

To fix this, we just need to set up threading instrumentation.

Note that this is a necessary but not sufficient change to get spans from the RPC rule validation call. I plan to get tracing working for RPC calls in a separate PR.

synced from Pro bf8845a7cdd39c38fb4780c6d3da80c32f280bbf